+++ 09_Texturing/Raytracer.cpp
@@ -24,7 +24,7 @@ void Raytracer::render(Image& img)
             Vec3 color;
             if (numSamples == 1)
             {
-                Ray r = computeRay(x, y, rs);
+                Ray r = computeRay(float(x), float(y), rs);
                 color = traceRay(r);
             }
             else
@@ -37,7 +37,7 @@ void Raytracer::render(Image& img)
                         color = color + traceRay(r);
                     }
                 }
-                color = color / numSamples;
+                color = color / float(numSamples);
             }
             img.setNormalizedValue(x, y, 0, color.r);
             img.setNormalizedValue(x, y, 1, color.g);
@@ -65,7 +65,7 @@ RaySetup Raytracer::computeRaySetup(const Image& img)
 
     Vec3 forwardDir = camera.getViewDir();
     Vec3 upDir = camera.getUpDir();
-    float openingAngle = camera.getFoV() * M_PI/180.0f;
+    float openingAngle = float(camera.getFoV() * M_PI/180.0);
     rs.rayOrigin = camera.getEyePoint();
 
     float aspectRatio = ((float)img.width) / ((float)img.height);
